Add codes_40x exceptions to endpoints
