+++ aria/api/responses.py
@@ -1,11 +1,1 @@
-from ninja import Schema
-
-
-class GenericResponse(Schema):
-    message: str
-    data: Schema
-
-
-class ExceptionResponse(Schema):
-    message: str
-    extra: dict[str, str] = {}
+codes_40x = frozenset({400, 401, 403, 404})
