+++ aria/api_auth/endpoints/public.py
@@ -3,7 +3,7 @@ from django.utils.translation import gettext as _
 from ninja import Router
 
 from aria.api.decorators import api
-from aria.api.responses import GenericResponse
+from aria.api.schemas.responses import GenericResponse
 from aria.api_auth.schemas.inputs import (
     TokenBlacklistInput,
     TokensObtainInput,
